refactor: move types around to avoid circular dependency
