+++ packages/analytics-js/src/components/capabilitiesManager/types.ts
@@ -1,6 +1,8 @@
 import { ILogger } from '@rudderstack/analytics-js/services/Logger/types';
-import { IErrorHandler } from '@rudderstack/analytics-js/services/ErrorHandler/types';
-import { IExternalSrcLoader } from '@rudderstack/analytics-js/services/ExternalSrcLoader/types';
+import {
+  IErrorHandler,
+  IExternalSrcLoader,
+} from '@rudderstack/analytics-js/services/ErrorHandler/types';
 
 export interface ICapabilitiesManager {
   logger?: ILogger;
