+++ packages/analytics-js/src/components/core/Analytics.ts
@@ -28,9 +28,11 @@ import {
 } from '@rudderstack/analytics-js/state/types';
 import { IHttpClient } from '@rudderstack/analytics-js/services/HttpClient/types';
 import { ILogger } from '@rudderstack/analytics-js/services/Logger/types';
-import { IErrorHandler } from '@rudderstack/analytics-js/services/ErrorHandler/types';
+import {
+  IErrorHandler,
+  IExternalSrcLoader,
+} from '@rudderstack/analytics-js/services/ErrorHandler/types';
 import { IPluginsManager } from '@rudderstack/analytics-js/components/pluginsManager/types';
-import { IExternalSrcLoader } from '@rudderstack/analytics-js/services/ExternalSrcLoader/types';
 import { IStoreManager } from '@rudderstack/analytics-js/services/StoreManager/types';
 import { IUserSessionManager } from '@rudderstack/analytics-js/components/userSessionManager/types';
 import { IConfigManager } from '@rudderstack/analytics-js/components/configManager/types';
