--- packages/analytics-js/src/services/ExternalSrcLoader/types.ts
@@ -1,19 +1,0 @@
-// eslint-disable-next-line import/no-cycle
-import { IErrorHandler } from '@rudderstack/analytics-js/services/ErrorHandler/types';
-import { ILogger } from '@rudderstack/analytics-js/services/Logger/types';
-
-export interface IExternalSourceLoadConfig {
-  url: string;
-  id: string;
-  callback?(id?: string): unknown;
-  async?: boolean;
-  timeout?: number;
-  extraAttributes?: Record<string, string>;
-}
-
-export interface IExternalSrcLoader {
-  errorHandler?: IErrorHandler;
-  logger?: ILogger;
-  timeout: number;
-  loadJSFile(config: IExternalSourceLoadConfig): Promise<void>;
-}
